+++ modules/string/src/cpp/StringContains.cpp
@@ -39,11 +39,13 @@ ArrayOf
 StringContains(ArrayOf A, ArrayOf Pattern, bool bCaseSensitive)
 {
     ArrayOf res;
-    if (A.isCharacterArray() && Pattern.isCharacterArray()) {
+    if ((A.isCharacterArray() || (A.isStringArray() && A.isScalar()))
+        && (Pattern.isCharacterArray() || (Pattern.isStringArray() && Pattern.isScalar()))) {
         res = ArrayOf::logicalConstructor(containsString(
             A.getContentAsWideString(), Pattern.getContentAsWideString(), bCaseSensitive));
     } else {
-        if (A.isCharacterArray() && IsCellOfString(Pattern)) {
+        if ((A.isCharacterArray() || (A.isStringArray() && A.isScalar())) &&
+			(Pattern.isStringArray() || IsCellOfString(Pattern))) {
             std::wstring strA = A.getContentAsWideString();
             Dimensions dimPattern = Pattern.getDimensions();
             size_t nbPattern = dimPattern.getElementCount();
@@ -57,7 +59,8 @@ StringContains(ArrayOf A, ArrayOf Pattern, bool bCaseSensitive)
                 }
             }
             res = ArrayOf::logicalConstructor(val);
-        } else if (IsCellOfString(A) && Pattern.isCharacterArray()) {
+        } else if ((A.isStringArray() || IsCellOfString(A)) &&
+			((Pattern.isStringArray() && Pattern.isScalar()) || Pattern.isCharacterArray())) {
             std::wstring pattern = Pattern.getContentAsWideString();
             Dimensions dimA = A.getDimensions();
             size_t nbA = dimA.getElementCount();
@@ -68,7 +71,7 @@ StringContains(ArrayOf A, ArrayOf Pattern, bool bCaseSensitive)
                     = containsString(cellA[k].getContentAsWideString(), pattern, bCaseSensitive);
             }
             res = ArrayOf(NLS_LOGICAL, dimA, result);
-        } else if (IsCellOfString(A) && IsCellOfString(Pattern)) {
+        } else if ((A.isStringArray() || IsCellOfString(A)) && (Pattern.isStringArray() || IsCellOfString(Pattern))) {
             Dimensions dimA = A.getDimensions();
             size_t nbA = dimA.getElementCount();
             Dimensions dimPattern = Pattern.getDimensions();